bug fix: wrong type used for line number representation

+++ microcontroller/src-atmel/borg/borgware-2d/games/tetris/view.c
@@ -295,7 +295,7 @@ void tetris_view_blinkLines(tetris_playfield_t *pPl)
 void tetris_view_showLineNumbers(tetris_view_t *pV)
 {
 	// get number of completed lines
-	uint8_t nLines = tetris_logic_getLines(pV->pLogic);
+	uint16_t nLines = tetris_logic_getLines(pV->pLogic);
 
 	// get decimal places
 	int8_t nOnes = nLines % 10;
